Move code outside of main
