+++ cmd/process-agent/main_common.go
@@ -8,6 +8,8 @@ package main
 import (
 	"context"
 	"fmt"
+	"github.com/DataDog/datadog-agent/pkg/process/runner"
+	"github.com/DataDog/datadog-agent/pkg/process/status"
 	"net/http"
 	_ "net/http/pprof"
 	"os"
@@ -169,7 +171,7 @@ func runAgent(globalParams *command.GlobalParams, exit chan struct{}) {
 	}
 	// we shouldn't quit because docker is not required. If no docker docket is available,
 	// we just pass down empty string
-	updateDockerSocket(dockerSock)
+	status.UpdateDockerSocket(dockerSock)
 
 	// use `internal_profiling.enabled` field in `process_config` section to enable/disable profiling for process-agent,
 	// but use the configuration from main agent to fill the settings
@@ -220,7 +222,11 @@ func runAgent(globalParams *command.GlobalParams, exit chan struct{}) {
 		return
 	}
 
-	err = initInfo(hostInfo.HostName, processModuleEnabled)
+	eps, err := runner.GetAPIEndpoints()
+	if err != nil {
+		log.Criticalf("Failed to initalize Api Endpoints: %s", err.Error())
+	}
+	err = status.InitInfo(hostInfo.HostName, processModuleEnabled, eps)
 	if err != nil {
 		log.Criticalf("Error initializing info: %s", err)
 		cleanupAndExit(1)
@@ -229,7 +235,7 @@ func runAgent(globalParams *command.GlobalParams, exit chan struct{}) {
 	if globalParams.Info {
 		// using the debug port to get info to work
 		url := fmt.Sprintf("http://localhost:%d/debug/vars", expVarPort)
-		if err := Info(os.Stdout, url); err != nil {
+		if err := status.Info(os.Stdout, url); err != nil {
 			cleanupAndExit(1)
 		}
 		return
@@ -252,20 +258,20 @@ func runAgent(globalParams *command.GlobalParams, exit chan struct{}) {
 		_ = log.Error(err)
 	}
 
-	cl, err := NewCollector(syscfg, hostInfo, enabledChecks)
+	cl, err := runner.NewCollector(syscfg, hostInfo, enabledChecks)
 	if err != nil {
 		log.Criticalf("Error creating collector: %s", err)
 		cleanupAndExit(1)
 		return
 	}
-	cl.submitter, err = NewSubmitter(hostInfo.HostName, cl.UpdateRTStatus)
+	cl.Submitter, err = runner.NewSubmitter(hostInfo.HostName, cl.UpdateRTStatus)
 	if err != nil {
 		log.Criticalf("Error creating checkSubmitter: %s", err)
 		cleanupAndExit(1)
 		return
 	}
 
-	if err := cl.run(exit); err != nil {
+	if err := cl.Run(exit); err != nil {
 		log.Criticalf("Error starting collector: %s", err)
 		os.Exit(1)
 		return
