+++ pkg/process/runner/collector_test.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-package main
+package runner
 
 import (
 	"testing"
@@ -16,7 +16,6 @@ import (
 	model "github.com/DataDog/agent-payload/v5/process"
 
 	processmocks "github.com/DataDog/datadog-agent/cmd/process-agent/mocks"
-	sysconfig "github.com/DataDog/datadog-agent/cmd/system-probe/config"
 	ddconfig "github.com/DataDog/datadog-agent/pkg/config"
 	"github.com/DataDog/datadog-agent/pkg/process/checks"
 	checkmocks "github.com/DataDog/datadog-agent/pkg/process/checks/mocks"
@@ -102,43 +101,6 @@ func TestHasContainers(t *testing.T) {
 	assert.Equal(1, getContainerCount(&collectorContainer))
 	assert.Equal(2, getContainerCount(&collectorRealTime))
 	assert.Equal(1, getContainerCount(&collectorContainerRealTime))
-}
-
-func TestDisableRealTime(t *testing.T) {
-	tests := []struct {
-		name            string
-		disableRealtime bool
-		expectedChecks  []string
-	}{
-		{
-			name:            "true",
-			disableRealtime: true,
-			expectedChecks:  []string{checks.ContainerCheckName},
-		},
-		{
-			name:            "false",
-			disableRealtime: false,
-			expectedChecks:  []string{checks.ContainerCheckName, checks.RTContainerCheckName},
-		},
-	}
-
-	assert := assert.New(t)
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			mockConfig := ddconfig.Mock(t)
-			mockConfig.Set("process_config.disable_realtime_checks", tc.disableRealtime)
-			mockConfig.Set("process_config.process_discovery.enabled", false) // Not an RT check so we don't care
-
-			enabledChecks := getChecks(&sysconfig.Config{}, true)
-			assert.EqualValues(tc.expectedChecks, getCheckNames(enabledChecks))
-
-			c, err := NewCollector(nil, &checks.HostInfo{}, enabledChecks)
-			assert.NoError(err)
-			assert.Equal(!tc.disableRealtime, c.runRealTime)
-			assert.ElementsMatch(tc.expectedChecks, getCheckNames(c.enabledChecks))
-		})
-	}
 }
 
 func TestDisableRealTimeProcessCheck(t *testing.T) {
@@ -243,7 +205,7 @@ func TestCollectorRunCheck(t *testing.T) {
 	require.NoError(t, err)
 	submitter := processmocks.NewSubmitter(t)
 	require.NoError(t, err)
-	c.submitter = submitter
+	c.Submitter = submitter
 
 	result := checks.StandardRunResult([]model.MessageBody{
 		&model.CollectorProc{},
